requirements.txt
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.3
Cython==3.0.12
dl==0.1.0
docutils==0.21.2
filelock==3.18.0
gradio==5.33.0
h5py==3.14.0
HTMLParser==0.0.2
ipython==8.21.0
ipython==8.12.3
ipywidgets==8.1.6
Jinja2==3.1.6
jnius==1.1.0
keyring==25.6.0
matplotlib==3.7.3
mock==5.2.0
numpy==1.25.2
opencv_python==4.11.0.86
pandas==2.0.3
Pillow==11.2.1
protobuf==3.20.3
pyOpenSSL==25.0.0
pyOpenSSL==25.1.0
railroad==0.5.0
redis==6.2.0
scikit_learn==1.7.0
scipy==1.15.3
Sphinx==8.2.3
tabulate==0.9.0
thread==2.0.5
torch==2.4.0+cu124
torch==2.7.1
torchvision==0.19.0+cu124
torchvision==0.22.1
tornado==6.4.2
tornado==6.5.1
tqdm==4.67.1
trimesh==4.6.11
trove_classifiers==2025.5.9.12
urllib3_secure_extra==0.1.0
wandb==0.20.1
xmlrpclib==1.0.1
